Extract remaining device command handlers
