+++ main.go
@@ -1,18 +1,15 @@
 package main
 
 import (
-	"bufio"
 	"bytes"
 	"context"
 	"crypto/x509"
 	"encoding/json"
 	"encoding/pem"
 	"fmt"
-	"io"
 	"log/slog"
 	"os"
 	"os/signal"
-	"path"
 	"path/filepath"
 	"runtime/debug"
 	"sort"
@@ -26,18 +23,8 @@ import (
 
 	"github.com/danielpaulus/go-ios/internal/clihelp"
 	"github.com/danielpaulus/go-ios/ios/debugproxy"
-	"github.com/danielpaulus/go-ios/ios/house_arrest"
 	"github.com/danielpaulus/go-ios/ios/tunnel"
 
-	"github.com/danielpaulus/go-ios/ios/amfi"
-
-	"github.com/danielpaulus/go-ios/ios/afc"
-	"github.com/danielpaulus/go-ios/ios/fileservice"
-
-	"github.com/danielpaulus/go-ios/ios/crashreport"
-	"github.com/danielpaulus/go-ios/ios/testmanagerd"
-
-	"github.com/danielpaulus/go-ios/ios/debugserver"
 	"github.com/danielpaulus/go-ios/ios/imagemounter"
 	"github.com/danielpaulus/go-ios/ios/zipconduit"
 
@@ -623,898 +610,6 @@ The commands work as following:
 		return
 	}
 
-	b, _ = arguments.Bool("erase")
-	if b {
-		force, _ := arguments.Bool("--force")
-		if !force {
-			slog.Warn("are you sure you want to erase device? (y/n)", "udid", device.Properties.SerialNumber)
-			reader := bufio.NewReader(os.Stdin)
-			// ReadString will block until the delimiter is entered
-			input, err := reader.ReadString('\n')
-			exitIfError("An error occured while reading input", err)
-			if !strings.HasPrefix(input, "y") {
-				slog.Error("abort")
-				return
-			}
-		}
-
-		exitIfError("failed erasing", mcinstall.Erase(device))
-		fmt.Print(convertToJSONString("ok"))
-		return
-	}
-
-	rsdCommand, _ := arguments.Bool("rsd")
-	if rsdCommand {
-		listCommand, _ := arguments.Bool("ls")
-		if listCommand {
-			services := device.Rsd.GetServices()
-			if JSONdisabled {
-				fmt.Println(services)
-			} else {
-				b, err := marshalJSON(services)
-				exitIfError("failed json conversion", err)
-				fmt.Println(string(b))
-			}
-			return
-		}
-	}
-
-	if mobileGestaltCommand(device, arguments) {
-		return
-	}
-
-	if deviceStateCommand {
-		if listCommand {
-			deviceState(device, true, false, "", "")
-			return
-		}
-		enable, _ := arguments.Bool("enable")
-		profileTypeId, _ := arguments.String("<profileTypeId>")
-		profileId, _ := arguments.String("<profileId>")
-		deviceState(device, false, enable, profileTypeId, profileId)
-	}
-
-	b, _ = arguments.Bool("wifi")
-	if b {
-		ssid, _ := arguments.String("--ssid")
-		psw, _ := arguments.String("--password")
-		encType, _ := arguments.String("--enc-type")
-		remove, _ := arguments.Bool("--remove")
-
-		if encType == "" {
-			encType = "WPA"
-		}
-
-		if remove {
-			exitIfError("failed removing wifi", mcinstall.RemoveWifi(device, ssid))
-		} else {
-			exitIfError("failed preparing wifi", mcinstall.PrepareWifi(device, ssid, psw, encType))
-		}
-		fmt.Print(convertToJSONString("ok"))
-		return
-	}
-
-	b, _ = arguments.Bool("prepare")
-	if b {
-		b, _ = arguments.Bool("create-cert")
-		if b {
-			cert, err := ios.CreateDERFormattedSupervisionCert()
-			exitIfError("failed creating cert", err)
-			err = os.WriteFile("supervision-cert.der", cert.CertDER, 0o777)
-			slog.Info("supervision-cert.der")
-			exitIfError("failed writing cert", err)
-			err = os.WriteFile("supervision-cert.pem", cert.CertPEM, 0o777)
-			slog.Info("supervision-cert.pem")
-			exitIfError("failed writing cert", err)
-			err = os.WriteFile("supervision-private-key.key", cert.PrivateKeyDER, 0o777)
-			slog.Info("supervision-private-key.key")
-			exitIfError("failed writing cert", err)
-			err = os.WriteFile("supervision-private-key.pem", cert.PrivateKeyPEM, 0o777)
-			slog.Info("supervision-private-key.pem")
-			exitIfError("failed writing key", err)
-			err = os.WriteFile("supervision-csr.csr", []byte(cert.Csr), 0o777)
-			slog.Info("supervision-csr.csr")
-			exitIfError("failed writing cert", err)
-			slog.Info("Golang does not have good PKCS12 format sadly. If you need a p12 file run this: " +
-				"'openssl pkcs12 -export -inkey supervision-private-key.pem -in supervision-cert.pem -out certificate.p12 -password pass:a'")
-			return
-		}
-		b, _ = arguments.Bool("printskip")
-		if b {
-			fmt.Println(convertToJSONString(mcinstall.GetAllSetupSkipOptions()))
-			return
-		}
-		b, _ = arguments.Bool("cloudconfig")
-		if b {
-			conn, err := mcinstall.New(device)
-			exitIfError("failed connecting to mcinstall", err)
-			defer conn.Close()
-			config, err := conn.GetCloudConfiguration()
-			exitIfError("failed getting cloud configuration", err)
-			fmt.Println(convertToJSONString(config))
-			return
-		}
-		skip := mcinstall.GetAllSetupSkipOptions()
-		skip1 := arguments["--skip"].([]string)
-		if len(skip1) > 0 {
-			skip = skip1
-		}
-
-		certfile, _ := arguments.String("--certfile")
-		orgname, _ := arguments.String("--orgname")
-		locale, _ := arguments.String("--locale")
-		lang, _ := arguments.String("--lang")
-		p12password, _ := arguments.String("--p12password")
-		if p12password == "" {
-			p12password = os.Getenv("P12_PASSWORD")
-		}
-		var certBytes []byte
-		if certfile != "" {
-			rawCertBytes, err := os.ReadFile(certfile)
-			exitIfError("failed opening cert file", err)
-			if orgname == "" {
-				logFatal("--orgname must be specified if certfile for supervision is provided")
-			}
-			certBytes, err = extractDERCertificate(rawCertBytes, p12password)
-			exitIfError("failed to parse supervision certificate", err)
-		}
-		exitIfError("failed erasing", mcinstall.Prepare(device, skip, certBytes, orgname, locale, lang))
-		fmt.Print(convertToJSONString("ok"))
-		return
-	}
-
-	b, _ = arguments.Bool("set-wallpaper")
-	if b {
-		imagePath, _ := arguments.String("<imagePath>")
-		p12file, _ := arguments.String("--p12file")
-		p12password, _ := arguments.String("--password")
-		if p12password == "" {
-			p12password = os.Getenv("P12_PASSWORD")
-		}
-		screen, _ := arguments.String("--screen")
-		if screen == "" {
-			screen = "home"
-		}
-		handleSetWallpaper(device, imagePath, screen, p12file, p12password)
-		return
-	}
-
-	b, _ = arguments.Bool("get-wallpaper")
-	if b {
-		out, _ := arguments.String("--output")
-		if out == "" {
-			out = "wallpaper.png"
-		}
-		handleGetWallpaper(device, out)
-		return
-	}
-
-	b, _ = arguments.Bool("get-icon-layout")
-	if b {
-		out, _ := arguments.String("--output")
-		handleGetIconLayout(device, out)
-		return
-	}
-
-	b, _ = arguments.Bool("set-icon-layout")
-	if b {
-		layoutFile, _ := arguments.String("<layoutFile>")
-		handleSetIconLayout(device, layoutFile)
-		return
-	}
-
-	if crashCommand(device, arguments) {
-		return
-	}
-	if instrumentsCommand(device, arguments) {
-		return
-	}
-
-	if imageCommand1(device, arguments) {
-		return
-	}
-
-	b, _ = arguments.Bool("assistivetouch")
-	if b {
-		force, _ := arguments.Bool("--force")
-		b, _ = arguments.Bool("enable")
-		if b {
-			assistiveTouch(device, "enable", force)
-		}
-		b, _ = arguments.Bool("disable")
-		if b {
-			assistiveTouch(device, "disable", force)
-		}
-		b, _ = arguments.Bool("toggle")
-		if b {
-			assistiveTouch(device, "toggle", force)
-		}
-		b, _ = arguments.Bool("get")
-		if b {
-			assistiveTouch(device, "get", force)
-		}
-	}
-
-	b, _ = arguments.Bool("voiceover")
-	if b {
-		force, _ := arguments.Bool("--force")
-		b, _ = arguments.Bool("enable")
-		if b {
-			voiceOver(device, "enable", force)
-		}
-		b, _ = arguments.Bool("disable")
-		if b {
-			voiceOver(device, "disable", force)
-		}
-		b, _ = arguments.Bool("toggle")
-		if b {
-			voiceOver(device, "toggle", force)
-		}
-		b, _ = arguments.Bool("get")
-		if b {
-			voiceOver(device, "get", force)
-		}
-	}
-
-	b, _ = arguments.Bool("zoom")
-	if b {
-		force, _ := arguments.Bool("--force")
-		b, _ = arguments.Bool("enable")
-		if b {
-			zoomTouch(device, "enable", force)
-		}
-		b, _ = arguments.Bool("disable")
-		if b {
-			zoomTouch(device, "disable", force)
-		}
-		b, _ = arguments.Bool("toggle")
-		if b {
-			zoomTouch(device, "toggle", force)
-		}
-		b, _ = arguments.Bool("get")
-		if b {
-			zoomTouch(device, "get", force)
-		}
-	}
-
-	lockdownCommand, _ := arguments.Bool("lockdown")
-	if lockdownCommand {
-		b, _ = arguments.Bool("get")
-		if b {
-			key := ""
-			if keyArg := arguments["<key>"]; keyArg != nil {
-				if keys, ok := keyArg.([]string); ok && len(keys) > 0 {
-					key = keys[0]
-				}
-			}
-			domain, _ := arguments.String("--domain")
-
-			lockdownConnection, err := ios.ConnectLockdownWithSession(device)
-			exitIfError("failed connecting to lockdown", err)
-			defer lockdownConnection.Close()
-
-			if key == "" && domain == "" {
-				// No key or domain specified, return all values
-				allValues, err := lockdownConnection.GetValues()
-				exitIfError("failed getting lockdown values", err)
-				fmt.Println(convertToJSONString(allValues.Value))
-			} else if domain != "" {
-				// Query from specific domain (key is optional, empty key returns all domain values)
-				value, err := lockdownConnection.GetValueForDomain(key, domain)
-				exitIfError(fmt.Sprintf("failed getting value from domain '%s'", domain), err)
-				fmt.Println(convertToJSONString(value))
-			} else {
-				// Query specific key from default domain
-				value, err := lockdownConnection.GetValue(key)
-				exitIfError(fmt.Sprintf("failed getting lockdown value '%s'", key), err)
-				fmt.Println(convertToJSONString(value))
-			}
-			return
-		}
-	}
-
-	b, _ = arguments.Bool("setlocation")
-	if b {
-		lat, _ := arguments.String("--lat")
-		lon, _ := arguments.String("--lon")
-
-		if device.SupportsRsd() {
-			server, err := instruments.NewLocationSimulationService(device)
-			exitIfError("failed to create location simulation service:", err)
-
-			startLocationSimulation(server, lat, lon)
-			return
-		}
-
-		setLocation(device, lat, lon)
-		return
-	}
-
-	b, _ = arguments.Bool("setlocationgpx")
-	if b {
-		gpxFilePath, _ := arguments.String("--gpxfilepath")
-		setLocationGPX(device, gpxFilePath)
-		return
-	}
-
-	b, _ = arguments.Bool("timeformat")
-	if b {
-		force, _ := arguments.Bool("--force")
-		b, _ = arguments.Bool("24h")
-		if b {
-			timeFormat(device, "24h", force)
-		}
-		b, _ = arguments.Bool("12h")
-		if b {
-			timeFormat(device, "12h", force)
-		}
-		b, _ = arguments.Bool("toggle")
-		if b {
-			timeFormat(device, "toggle", force)
-		}
-		b, _ = arguments.Bool("get")
-		if b {
-			timeFormat(device, "get", force)
-		}
-	}
-
-	b, _ = arguments.Bool("httpproxy")
-	if b {
-		removeCommand, _ := arguments.Bool("remove")
-		if removeCommand {
-			mcinstall.RemoveProxy(device)
-			exitIfError("failed removing proxy", err)
-			slog.Info("success")
-			return
-		}
-		host, _ := arguments.String("<host>")
-		port, _ := arguments.String("<port>")
-		user, _ := arguments.String("<user>")
-		pass, _ := arguments.String("<pass>")
-		if pass == "" {
-			pass = os.Getenv("PROXY_PASSWORD")
-		}
-		p12file, _ := arguments.String("--p12file")
-		p12password, _ := arguments.String("--password")
-		if p12password == "" {
-			p12password = os.Getenv("P12_PASSWORD")
-		}
-		p12bytes, err := os.ReadFile(p12file)
-		exitIfError("could not read p12-file", err)
-
-		err = mcinstall.SetHttpProxy(device, host, port, user, pass, p12bytes, p12password)
-		exitIfError("failed", err)
-		slog.Info("success")
-		return
-	}
-
-	b, _ = arguments.Bool("profile")
-	if b {
-		if listCommand {
-			handleProfileList(device)
-		}
-		b, _ = arguments.Bool("add")
-		if b {
-			name, _ := arguments.String("<profileFile>")
-			p12file, _ := arguments.String("--p12file")
-			p12password, _ := arguments.String("--password")
-			if p12password == "" {
-				p12password = os.Getenv("P12_PASSWORD")
-			}
-			if p12file != "" {
-				handleProfileAddSupervised(device, name, p12file, p12password)
-				return
-			}
-			handleProfileAdd(device, name)
-		}
-		b, _ = arguments.Bool("remove")
-		if b {
-			name, _ := arguments.String("<profileName>")
-			handleProfileRemove(device, name)
-		}
-
-		return
-	}
-
-	b, _ = arguments.Bool("forward")
-	if b {
-		// Check for new --port syntax first (multi-forward)
-		mappings, _ := arguments["--port"].([]string)
-		if len(mappings) > 0 {
-			startMultiForwarding(device, mappings)
-		} else {
-			// Backwards compatible: single forward
-			hostPort, _ := arguments.Int("<hostPort>")
-			targetPort, _ := arguments.Int("<targetPort>")
-			startForwarding(device, uint16(hostPort), uint16(targetPort))
-		}
-		return
-	}
-
-	b, _ = arguments.Bool("launch")
-	if b {
-		wait, _ := arguments.Bool("--wait")
-		bKillExisting, _ := arguments.Bool("--kill-existing")
-		bundleID, _ := arguments.String("<bundleID>")
-		if bundleID == "" {
-			logFatal("please provide a bundleID")
-		}
-		pControl, err := instruments.NewProcessControl(device)
-		exitIfError("processcontrol failed", err)
-		opts := map[string]any{}
-		if bKillExisting {
-			opts["KillExisting"] = 1
-		} // end if
-		args := toArgs(arguments["--arg"].([]string))
-		envs := toEnvs(arguments["--env"].([]string))
-		pid, err := pControl.LaunchAppWithArgs(bundleID, args, envs, opts)
-		exitIfError("launch app command failed", err)
-		slog.Info("Process launched", "pid", pid)
-		if wait {
-			c := make(chan os.Signal, 1)
-			signal.Notify(c, syscall.SIGINT, syscall.SIGTERM)
-			<-c
-			slog.Info("stop listening to logs", "pid", pid)
-		}
-	}
-
-	b, _ = arguments.Bool("sysmontap")
-	if b {
-		printSysmontapStats(device)
-	}
-
-	b, _ = arguments.Bool("memlimitoff")
-	if b {
-		processName, _ := arguments.String("--process")
-
-		pControl, err := instruments.NewProcessControl(device)
-		exitIfError("processcontrol failed", err)
-		defer pControl.Close()
-
-		svc, err := instruments.NewDeviceInfoService(device)
-		exitIfError("failed opening deviceInfoService for getting process list", err)
-		defer svc.Close()
-
-		process, err := svc.ProcessByName(processName)
-		exitIfError("process not found", err)
-		if process.Pid > 1 {
-			disabled, err := pControl.DisableMemoryLimit(process.Pid)
-			exitIfError("DisableMemoryLimit failed", err)
-			slog.Info("memory limit is off", "process", process.Name, "pid", process.Pid, "disabled", disabled)
-		}
-	}
-
-	b, _ = arguments.Bool("kill")
-	if b {
-		var response []installationproxy.AppInfo
-		bundleID, _ := arguments.String("<bundleID>")
-		processIDint, _ := arguments.Int("--pid")
-		processName, _ := arguments.String("--process")
-
-		processID := uint64(processIDint)
-
-		// Technically "Mach Kernel" is process 0, I suppose we provide no way to attempt to kill that.
-		if bundleID == "" && processID == 0 && processName == "" {
-			logFatal("please provide a bundleID")
-		}
-		pControl, err := instruments.NewProcessControl(device)
-		exitIfError("processcontrol failed", err)
-		svc, _ := installationproxy.New(device)
-
-		// Look for correct process exe name for this bundleID. By default, searches only user-installed apps.
-		if bundleID != "" {
-			response, err = svc.BrowseAllApps()
-			exitIfError("browsing apps failed", err)
-
-			for _, app := range response {
-				if app.CFBundleIdentifier() == bundleID {
-					processName = app.CFBundleExecutable()
-					break
-				}
-			}
-			if processName == "" {
-				slog.Error("not installed", "bundleID", bundleID)
-				os.Exit(1)
-				return
-			}
-		}
-
-		service, err := instruments.NewDeviceInfoService(device)
-		defer service.Close()
-		exitIfError("failed opening deviceInfoService for getting process list", err)
-		processList, _ := service.ProcessList()
-		// ps
-		for _, p := range processList {
-			if (processID > 0 && p.Pid == processID) || (processName != "" && p.Name == processName) {
-				err = pControl.KillProcess(p.Pid)
-				exitIfError("kill process failed ", err)
-				if bundleID != "" {
-					slog.Info("killed", "bundleID", bundleID, "pid", p.Pid)
-				} else {
-					slog.Info("killed", "process", p.Name, "pid", p.Pid)
-				}
-				return
-			}
-		}
-		if bundleID != "" {
-			slog.Error("process not found", "bundleID", bundleID)
-		} else if processName != "" {
-			slog.Error("process not found", "process", processName)
-		} else {
-			slog.Error("process not found", "pid", processID)
-		}
-		os.Exit(1)
-		return
-	}
-
-	b, _ = arguments.Bool("runtest")
-	if b {
-		bundleID, _ := arguments.String("--bundle-id")
-		testRunnerBundleId, _ := arguments.String("--test-runner-bundle-id")
-		xctestConfig, _ := arguments.String("--xctest-config")
-
-		testsToRunArg := arguments["--test-to-run"]
-		var testsToRun []string
-		if testsToRunArg != nil && len(testsToRunArg.([]string)) > 0 {
-			testsToRun = testsToRunArg.([]string)
-		}
-
-		testsToSkipArg := arguments["--test-to-skip"]
-		var testsToSkip []string
-		testsToSkip = nil
-		if testsToSkipArg != nil && len(testsToSkipArg.([]string)) > 0 {
-			testsToSkip = testsToSkipArg.([]string)
-		}
-
-		rawTestlog, rawTestlogErr := arguments.String("--log-output")
-		env := splitKeyValuePairs(arguments["--env"].([]string), "=")
-		isXCTest, _ := arguments.Bool("--xctest")
-
-		config := testmanagerd.TestConfig{
-			BundleId:           bundleID,
-			TestRunnerBundleId: testRunnerBundleId,
-			XctestConfigName:   xctestConfig,
-			Env:                env,
-			TestsToRun:         testsToRun,
-			TestsToSkip:        testsToSkip,
-			XcTest:             isXCTest,
-			Device:             device,
-		}
-
-		if rawTestlogErr == nil {
-			var writer *os.File = os.Stdout
-			if rawTestlog != "-" {
-				file, err := os.Create(rawTestlog)
-				exitIfError("Cannot open file "+rawTestlog, err)
-				writer = file
-			}
-			defer writer.Close()
-
-			config.Listener = testmanagerd.NewTestListener(writer, writer, os.TempDir())
-
-			testResults, err := testmanagerd.RunTestWithConfig(context.TODO(), config)
-			if err != nil {
-				slog.Info("Failed running Xcuitest", "error", err)
-			}
-
-			slog.Info("test results", "results", testResults)
-		} else {
-			config.Listener = testmanagerd.NewTestListener(io.Discard, io.Discard, os.TempDir())
-			_, err := testmanagerd.RunTestWithConfig(context.TODO(), config)
-			if err != nil {
-				slog.Info("Failed running Xcuitest", "error", err)
-			}
-		}
-		return
-	}
-
-	b, _ = arguments.Bool("runxctest")
-	if b {
-		xctestrunFilePath, _ := arguments.String("--xctestrun-file-path")
-
-		rawTestlog, rawTestlogErr := arguments.String("--log-output")
-
-		if rawTestlogErr == nil {
-			var writer *os.File = os.Stdout
-			if rawTestlog != "-" {
-				file, err := os.Create(rawTestlog)
-				exitIfError("Cannot open file "+rawTestlog, err)
-				writer = file
-			}
-			defer writer.Close()
-			var listener = testmanagerd.NewTestListener(writer, writer, os.TempDir())
-
-			testResults, err := testmanagerd.StartXCTestWithConfig(context.TODO(), xctestrunFilePath, device, listener)
-			if err != nil {
-				slog.Info("Failed running Xctest", "error", err)
-			}
-
-			slog.Info("test results", "results", testResults)
-		} else {
-			var listener = testmanagerd.NewTestListener(io.Discard, io.Discard, os.TempDir())
-			_, err := testmanagerd.StartXCTestWithConfig(context.TODO(), xctestrunFilePath, device, listener)
-			if err != nil {
-				slog.Info("Failed running Xctest", "error", err)
-			}
-		}
-		return
-	}
-
-	if runWdaCommand(device, arguments) {
-		return
-	}
-
-	b, _ = arguments.Bool("ax")
-	if b {
-		startAx(device, arguments)
-		return
-	}
-
-	b, _ = arguments.Bool("resetax")
-	if b {
-		resetAx(device)
-		return
-	}
-
-	b, _ = arguments.Bool("debug")
-	if b {
-		appPath, _ := arguments.String("<app_path>")
-		if appPath == "" {
-			logFatal("parameter bundleid and app_path must be specified")
-		}
-		stopAtEntry, _ := arguments.Bool("--stop-at-entry")
-		exitIfError("debug server failed", debugserver.Start(device, appPath, stopAtEntry))
-	}
-
-	b, _ = arguments.Bool("file")
-	if b {
-		// file command uses RemoteXPC (iOS 17+) and requires tunnel
-		if !device.SupportsRsd() {
-			exitIfError("file command requires iOS 17+ with tunnel", fmt.Errorf("tunnel not running. Start with: ios tunnel start"))
-		}
-
-		// Determine domain from flags
-		bundleID, _ := arguments.String("--app")
-		groupID, _ := arguments.String("--app-group")
-		useCrash, _ := arguments.Bool("--crash")
-		useTemp, _ := arguments.Bool("--temp")
-
-		// Count how many domain flags were specified
-		flagCount := 0
-		if bundleID != "" {
-			flagCount++
-		}
-		if groupID != "" {
-			flagCount++
-		}
-		if useCrash {
-			flagCount++
-		}
-		if useTemp {
-			flagCount++
-		}
-
-		if flagCount > 1 {
-			exitIfError("file command", fmt.Errorf("can only specify one of: --app, --app-group, --crash, or --temp"))
-		}
-		if flagCount == 0 {
-			exitIfError("file command", fmt.Errorf("must specify one of: --app=<bundleID>, --app-group=<groupID>, --crash, or --temp"))
-		}
-
-		// Determine domain and identifier
-		var domain fileservice.Domain
-		var identifier string
-
-		if bundleID != "" {
-			domain = fileservice.DomainAppDataContainer
-			identifier = bundleID
-		} else if groupID != "" {
-			domain = fileservice.DomainAppGroupDataContainer
-			identifier = groupID
-		} else if useCrash {
-			domain = fileservice.DomainSystemCrashLogs
-			identifier = ""
-		} else if useTemp {
-			domain = fileservice.DomainTemporary
-			identifier = ""
-		}
-
-		// Create connection
-		conn, err := fileservice.New(device, domain, identifier)
-		exitIfError("file: failed to connect to file service", err)
-		defer func() {
-			if closeErr := conn.Close(); closeErr != nil {
-				slog.Error("Failed to close file service connection", "error", closeErr)
-			}
-		}()
-
-		// Handle ls subcommand
-		b, _ = arguments.Bool("ls")
-		if b {
-			path, _ := arguments.String("--path")
-			if path == "" {
-				path = "."
-			}
-
-			files, err := conn.ListDirectory(path)
-			exitIfError("file ls: failed to list directory", err)
-
-			if !JSONdisabled {
-				result := map[string]interface{}{
-					"path":  path,
-					"files": files,
-					"count": len(files),
-				}
-				fmt.Println(convertToJSONString(result))
-			} else {
-				fmt.Printf("Files in %s:\n", path)
-				for _, file := range files {
-					fmt.Printf("  %s\n", file)
-				}
-				fmt.Printf("\nTotal: %d files\n", len(files))
-			}
-		}
-
-		// Handle pull subcommand
-		b, _ = arguments.Bool("pull")
-		if b {
-			remotePath, _ := arguments.String("--remote")
-			localPath, _ := arguments.String("--local")
-
-			if remotePath == "" {
-				exitIfError("file pull", fmt.Errorf("--remote=<path> is required"))
-			}
-			if localPath == "" {
-				exitIfError("file pull", fmt.Errorf("--local=<path> is required"))
-			}
-
-			// Create output file for streaming
-			outputFile, err := os.Create(localPath)
-			exitIfError("file pull: failed to create output file", err)
-			defer outputFile.Close()
-
-			// Download file using streaming to minimize memory usage
-			slog.Info(fmt.Sprintf("Downloading %s to %s...", remotePath, localPath))
-			err = conn.PullFile(remotePath, outputFile)
-			exitIfError("file pull: failed to download file", err)
-
-			// Get file size for reporting
-			fileInfo, err := outputFile.Stat()
-			exitIfError("file pull: failed to get file info", err)
-			fileSize := fileInfo.Size()
-
-			if !JSONdisabled {
-				result := map[string]interface{}{
-					"remote": remotePath,
-					"local":  localPath,
-					"size":   fileSize,
-				}
-				fmt.Println(convertToJSONString(result))
-			} else {
-				slog.Info(fmt.Sprintf("Downloaded %d bytes to %s", fileSize, localPath))
-			}
-		}
-
-		b, _ = arguments.Bool("push")
-		if b {
-			localPath, _ := arguments.String("--local")
-			remotePath, _ := arguments.String("--remote")
-
-			if localPath == "" || remotePath == "" {
-				exitIfError("push requires --local and --remote paths", fmt.Errorf("missing required arguments"))
-			}
-
-			// Get file info to preserve permissions
-			fileInfo, err := os.Stat(localPath)
-			exitIfError("push: failed to stat local file", err)
-
-			// Get file permissions from source file, default UID (501) and GID (501)
-			permissions := int64(fileInfo.Mode().Perm())
-			uid := int64(501)
-			gid := int64(501)
-			fileSize := fileInfo.Size()
-
-			// Open file for streaming
-			file, err := os.Open(localPath)
-			exitIfError("push: failed to open local file", err)
-			defer file.Close()
-
-			// Upload file using streaming to minimize memory usage
-			slog.Info(fmt.Sprintf("Uploading %s to %s...", localPath, remotePath))
-			err = conn.PushFile(remotePath, file, fileSize, permissions, uid, gid)
-			exitIfError("push: failed to upload file", err)
-
-			if !JSONdisabled {
-				result := map[string]interface{}{
-					"remote": remotePath,
-					"local":  localPath,
-					"size":   fileSize,
-				}
-				fmt.Println(convertToJSONString(result))
-			} else {
-				slog.Info(fmt.Sprintf("Uploaded %d bytes to %s", fileSize, remotePath))
-			}
-		}
-
-		return
-	}
-
-	b, _ = arguments.Bool("fsync")
-	if b {
-		containerBundleId, _ := arguments.String("--app")
-		var afcService *afc.Client
-		if containerBundleId == "" {
-			afcService, err = afc.New(device)
-		} else {
-			afcService, err = house_arrest.New(device, containerBundleId)
-		}
-		exitIfError("fsync: connect afc service failed", err)
-		b, _ = arguments.Bool("rm")
-		if b {
-			path, _ := arguments.String("--path")
-			isRecursive, _ := arguments.Bool("--r")
-			if isRecursive {
-				err = afcService.RemoveAll(path)
-			} else {
-				err = afcService.Remove(path)
-			}
-			exitIfError("fsync: remove failed", err)
-		}
-
-		b, _ = arguments.Bool("tree")
-		if b {
-			path, _ := arguments.String("--path")
-			err := afcService.WalkDir(path, func(path string, info afc.FileInfo, err error) error {
-				s := strings.Split(path, string(os.PathSeparator))
-				_, f := filepath.Split(path)
-				prefix := strings.Repeat("|  ", len(s)-1)
-
-				suffix := ""
-				if info.Type == afc.S_IFDIR {
-					suffix = "/"
-				}
-
-				fmt.Printf("%s|-%s%s\n", prefix, f, suffix)
-				return nil
-			})
-			exitIfError("fsync: tree view failed", err)
-		}
-
-		b, _ = arguments.Bool("mkdir")
-		if b {
-			path, _ := arguments.String("--path")
-			err = afcService.MkDir(path)
-			exitIfError("fsync: mkdir failed", err)
-		}
-
-		b, _ = arguments.Bool("pull")
-		if b {
-			sp, _ := arguments.String("--srcPath")
-			dp, _ := arguments.String("--dstPath")
-			if dp != "" {
-				ret, _ := ios.PathExists(dp)
-				if !ret {
-					err = os.MkdirAll(dp, os.ModePerm)
-					exitIfError("mkdir failed", err)
-				}
-			}
-
-			dp = path.Join(dp, filepath.Base(sp))
-			err = afcService.Pull(sp, dp)
-			exitIfError("fsync: pull failed", err)
-		}
-		b, _ = arguments.Bool("push")
-		if b {
-			sp, _ := arguments.String("--srcPath")
-			dp, _ := arguments.String("--dstPath")
-
-			err = afcService.Push(sp, dp)
-			exitIfError("fsync: push failed", err)
-		}
-		afcService.Close()
-		return
-	}
-
 	if tunnelCommand {
 		startCommand, _ := arguments.Bool("start")
 		useUserspaceNetworking, _ := arguments.Bool("--userspace")
@@ -1560,38 +655,6 @@ The commands work as following:
 		}
 	}
 
-	b, _ = arguments.Bool("devmode")
-	if b {
-		enable, _ := arguments.Bool("enable")
-		get, _ := arguments.Bool("get")
-		enablePostRestart, _ := arguments.Bool("--enable-post-restart")
-		if enable {
-			err := amfi.EnableDeveloperMode(device, enablePostRestart)
-			exitIfError("Failed enabling developer mode", err)
-		}
-
-		if get {
-			devModeEnabled, _ := imagemounter.IsDevModeEnabled(device)
-			if JSONdisabled {
-				fmt.Printf("Developer mode enabled: %v\n", devModeEnabled)
-			} else {
-				result := map[string]interface{}{"DeveloperModeEnabled": devModeEnabled}
-				fmt.Println(convertToJSONString(result))
-			}
-		}
-
-		reveal, _ := arguments.Bool("reveal")
-		if reveal {
-			conn, err := amfi.New(device)
-			exitIfError("Failed connecting to AMFI service", err)
-			defer conn.Close()
-			err = conn.RevealDevMode()
-			exitIfError("Failed revealing developer mode menu", err)
-			slog.Info("Developer Mode menu has been revealed on the device. Go to Settings → Privacy & Security → Developer Mode to enable it.")
-		}
-
-		return
-	}
 }
 
 func printSysmontapStats(device ios.DeviceEntry) {
@@ -1630,177 +693,6 @@ func printSysmontapStats(device ios.DeviceEntry) {
 	}
 }
 
-func mobileGestaltCommand(device ios.DeviceEntry, arguments docopt.Opts) bool {
-	b, _ := arguments.Bool("mobilegestalt")
-	if b {
-		conn, _ := diagnostics.New(device)
-		keys := arguments["<key>"].([]string)
-		plist, _ := arguments.Bool("--plist")
-		resp, _ := conn.MobileGestaltQuery(keys)
-		if plist {
-			fmt.Printf("%s\n", ios.ToPlist(resp))
-			return true
-		}
-		jb, _ := marshalJSON(resp)
-		fmt.Printf("%s\n", jb)
-		return true
-	}
-	return b
-}
-
-func imageCommand1(device ios.DeviceEntry, arguments docopt.Opts) bool {
-	b, _ := arguments.Bool("image")
-	if b {
-		list, _ := arguments.Bool("list")
-		if list {
-			listMountedImages(device)
-		}
-
-		path, _ := arguments.String("--path")
-
-		auto, _ := arguments.Bool("auto")
-		if auto {
-			basedir, _ := arguments.String("--basedir")
-			if basedir == "" {
-				basedir = "./devimages"
-			}
-
-			var err error
-			path, err = imagemounter.DownloadImageFor(device, basedir)
-			if err != nil {
-				slog.Error("failed downloading image", "basedir", basedir, "udid", device.Properties.SerialNumber, "err", err)
-				return false
-			}
-
-			slog.Info("success downloaded image", "basedir", basedir, "udid", device.Properties.SerialNumber)
-		}
-
-		mount, _ := arguments.Bool("mount")
-		if mount || auto {
-			err := imagemounter.MountImage(device, path)
-			if err != nil {
-				slog.Error("error mounting image", "image", path, "udid", device.Properties.SerialNumber, "err", err)
-				return true
-			}
-			slog.Info("success mounting image", "image", path, "udid", device.Properties.SerialNumber)
-		}
-
-		unmount, _ := arguments.Bool("unmount")
-		if unmount {
-			err := imagemounter.UnmountImage(device)
-			if err != nil {
-				slog.Error("error unmounting image", "udid", device.Properties.SerialNumber, "err", err)
-				return true
-			}
-			slog.Info("success unmounting image", "udid", device.Properties.SerialNumber)
-		}
-	}
-	return b
-}
-
-func runWdaCommand(device ios.DeviceEntry, arguments docopt.Opts) bool {
-	b, _ := arguments.Bool("runwda")
-	if b {
-		bundleID, _ := arguments.String("--bundleid")
-		testbundleID, _ := arguments.String("--testrunnerbundleid")
-		xctestconfig, _ := arguments.String("--xctestconfig")
-		wdaargs := arguments["--arg"].([]string)
-		wdaenv := splitKeyValuePairs(arguments["--env"].([]string), "=")
-
-		if bundleID == "" && testbundleID == "" && xctestconfig == "" {
-			slog.Info("no bundle ids specified, falling back to defaults")
-			bundleID, testbundleID, xctestconfig = "com.facebook.WebDriverAgentRunner.xctrunner", "com.facebook.WebDriverAgentRunner.xctrunner", "WebDriverAgentRunner.xctest"
-		}
-		if bundleID == "" || testbundleID == "" || xctestconfig == "" {
-			slog.Error("please specify either NONE of bundleid, testbundleid and xctestconfig or ALL of them. At least one was empty.", "bundleid", bundleID, "testbundleid", testbundleID, "xctestconfig", xctestconfig)
-			return true
-		}
-		slog.Info("Running wda", "bundleid", bundleID, "testbundleid", testbundleID, "xctestconfig", xctestconfig)
-
-		rawTestlog, rawTestlogErr := arguments.String("--log-output")
-
-		var writer io.Writer
-
-		if rawTestlogErr == nil {
-			writerCloser := os.Stdout
-			writer = writerCloser
-			if rawTestlog != "-" {
-				file, err := os.Create(rawTestlog)
-				exitIfError("Cannot open file "+rawTestlog, err)
-				writer = file
-			}
-			defer writerCloser.Close()
-		} else {
-			writer = io.Discard
-		}
-
-		errorChannel := make(chan error)
-		defer close(errorChannel)
-		ctx, stopWda := context.WithCancel(context.Background())
-		go func() {
-			_, err := testmanagerd.RunTestWithConfig(ctx, testmanagerd.TestConfig{
-				BundleId:           bundleID,
-				TestRunnerBundleId: testbundleID,
-				XctestConfigName:   xctestconfig,
-				Env:                wdaenv,
-				Args:               wdaargs,
-				Device:             device,
-				Listener:           testmanagerd.NewTestListener(writer, writer, os.TempDir()),
-			})
-			if err != nil {
-				errorChannel <- err
-			}
-			stopWda()
-		}()
-		c := make(chan os.Signal, 1)
-		signal.Notify(c, syscall.SIGINT, syscall.SIGTERM)
-
-		select {
-		case err := <-errorChannel:
-			slog.Error("Failed running WDA", "error", err)
-			stopWda()
-			os.Exit(1)
-		case <-ctx.Done():
-			slog.Error("WDA process ended unexpectedly")
-			os.Exit(1)
-		case signal := <-c:
-			slog.Info(fmt.Sprintf("os signal %d received, closing...", signal))
-			stopWda()
-		}
-		slog.Info("Done Closing")
-	}
-	return b
-}
-
-func instrumentsCommand(device ios.DeviceEntry, arguments docopt.Opts) bool {
-	b, _ := arguments.Bool("instruments")
-	if b {
-		listenerFunc, closeFunc, err := instruments.ListenAppStateNotifications(device)
-		if err != nil {
-			logFatal("failed listening to app state notifications", "error", err)
-		}
-		go func() {
-			for {
-				notification, err := listenerFunc()
-				if err != nil {
-					slog.Error("listener error", "error", err)
-					return
-				}
-				s, _ := json.Marshal(notification)
-				fmt.Println(string(s))
-			}
-		}()
-		c := make(chan os.Signal, 1)
-		signal.Notify(c, syscall.SIGINT, syscall.SIGTERM)
-		<-c
-		err = closeFunc()
-		if err != nil {
-			slog.Warn("timeout during close", "error", err)
-		}
-	}
-	return b
-}
-
 func toArgs(argsIn []string) []interface{} {
 	args := []interface{}{}
 	for _, arg := range argsIn {
@@ -1820,44 +712,6 @@ func toEnvs(envsIn []string) map[string]interface{} {
 	}
 
 	return env
-}
-
-func crashCommand(device ios.DeviceEntry, arguments docopt.Opts) bool {
-	b, _ := arguments.Bool("crash")
-	if b {
-		ls, _ := arguments.Bool("ls")
-		if ls {
-			pattern, err := arguments.String("<pattern>")
-			if err != nil || pattern == "" {
-				pattern = "*"
-			}
-			files, err := crashreport.ListReports(device, pattern)
-			exitIfError("failed listing crashreports", err)
-			fmt.Println(
-				convertToJSONString(
-					map[string]interface{}{"files": files, "length": len(files)},
-				),
-			)
-		}
-		cp, _ := arguments.Bool("cp")
-		if cp {
-			pattern, _ := arguments.String("<srcpattern>")
-			target, _ := arguments.String("<target>")
-			slog.Debug("cp", "srcpattern", pattern, "target", target)
-			err := crashreport.DownloadReports(device, pattern, target)
-			exitIfError("failed downloading crashreports", err)
-		}
-
-		rm, _ := arguments.Bool("rm")
-		if rm {
-			cwd, _ := arguments.String("<cwd>")
-			pattern, _ := arguments.String("<pattern>")
-			slog.Debug("rm", "cwd", cwd, "pattern", pattern)
-			err := crashreport.RemoveReports(device, cwd, pattern)
-			exitIfError("failed deleting crashreports", err)
-		}
-	}
-	return b
 }
 
 func deviceState(device ios.DeviceEntry, list bool, enable bool, profileTypeId string, profileId string) {
